requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
Jinja2==3.1.4
python-multipart==0.0.9
SQLAlchemy==2.0.30
pydantic==2.7.1
pillow==10.3.0
